modificada la logica cuando intenta limpiar un archivo inexistente

+++ Practicas/fork_fd_clase4/escritores.py
@@ -42,6 +42,11 @@ Proceso 401709 escribiendo letra 'C'
 ABCBACABCABCABC
 """
 import argparse, os, time
+from importlib.resources import path
+
+class ToBig(Exception):
+    def __init__(self,message):
+        print(message)
 
 #this method write to the file the letter
 def file_writter(path_dir,letter_of_this_process):
@@ -96,17 +101,22 @@ def main():
     parser.add_argument('-r', '--amount',metavar='AMOUNT', type=int,default=1,help='Amount of times that the letter repeats itself')
     parser.add_argument('-v', '--verbose',action='store_true', help="Activate verbose mode")
     letters = ['A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z']
+
     args = parser.parse_args()
     number = args.number
     path_dir = args.path
     amount = args.amount
     verbose = args.verbose
+    
+    if number > 26:
+        raise ToBig(f"cant assing one letter to every process letters:26 , process:{number} ")
     #this clean the file if there is any
-    open(path_dir, 'w').close()
+    if os.path.isfile(path_dir):
+        open(path_dir, 'w').close()
+    
+    
     creator(number,verbose,amount,path_dir,letters)
-    
 
-    
     
 if __name__ == "__main__":
     main()
